add pre-init check
