+++ src/k8s/pkg/k8sd/app/hooks_bootstrap.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"net"
 	"net/http"
-	"os"
 	"path/filepath"
 	"time"
 
@@ -22,46 +21,8 @@ import (
 	snaputil "github.com/canonical/k8s/pkg/snap/util"
 	"github.com/canonical/k8s/pkg/utils"
 	"github.com/canonical/k8s/pkg/utils/experimental/snapdconfig"
-	"github.com/canonical/lxd/shared"
-	microclusterTypes "github.com/canonical/microcluster/v3/rest/types"
 	"github.com/canonical/microcluster/v3/state"
 )
-
-// onPreInit is called before we bootstrap or join a node.
-func (a *App) onPreInit(ctx context.Context, s state.State, bootstrap bool, initConfig map[string]string) error {
-	err := os.Remove(filepath.Join(s.FileSystem().StateDir, "server.crt"))
-	if err != nil {
-		return fmt.Errorf("failed to remove server.crt: %w", err)
-	}
-
-	err = os.Remove(filepath.Join(s.FileSystem().StateDir, "server.key"))
-	if err != nil {
-		return fmt.Errorf("failed to remove server.key: %w", err)
-	}
-
-	cert, err := shared.KeyPairAndCA(
-		s.FileSystem().StateDir,
-		string(microclusterTypes.ServerCertificateName),
-		shared.CertServer,
-		shared.CertOptions{
-			AddHosts:                true,
-			CommonName:              s.Name(),
-			SubjectAlternativeNames: []string{initConfig["cluster_name"]},
-		})
-	if err != nil {
-		return err
-	}
-
-	err = a.client.UpdateCertificate(ctx, microclusterTypes.ServerCertificateName, microclusterTypes.KeyPair{
-		Cert: string(cert.PublicKey()),
-		Key:  string(cert.PrivateKey()),
-	})
-	if err != nil {
-		return fmt.Errorf("failed to update certificate %s: %w", microclusterTypes.ServerCertificateName, err)
-	}
-
-	return nil
-}
 
 // onBootstrap is called after we bootstrap the first cluster node.
 // onBootstrap configures local services then writes the cluster config on the database.
